requirements.txt
sounddevice
soundfile
typer
python-dotenv
elevenlabs
google-generativeai
